services/monitoring/requirements.txt
fastapi>=0.115.0
starlette>=0.37.0
uvicorn
httpx<0.28
pydantic>=2.4,<2.7
pydantic-settings>=2.0,<2.2
prometheus-client
opensearch-py
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
redis
PyYAML
pandas
psutil
